feature: create css module for Main page
