+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import './App.css';
-import Main from './pages/main';
+import Main from './pages/Main/main';
 import Login from './pages/login';
 import PlanList from './pages/plan-list';
 import Plans from './pages/plans';
